UML-32 moved LinkValidator
