+++ src/components/ComponentDiagram/elements/Node/Node.ts
@@ -4,7 +4,7 @@ import { Port, PortEvents } from '../Port/Port';
 import ComponentType from '../../../../models/ComponentType';
 import ComponentFactory from '../../../../models/factories/ComponentFactory';
 import { ComponentI } from '../../../../models/components/Component';
-import { LinkValidatorI } from '../LinkValidator';
+import { LinkValidatorI } from '../../LinkValidator';
 import Observable from '../../../../lib/Observable';
 
 export interface NodeI extends NodeModel {
