+++ src/components/ComponentDiagram/elements/Node/NodeInterface/NodeInterfaceFactory.tsx
@@ -7,7 +7,7 @@ import NodeInterfaceWidget from './NodeInterfaceWidget';
 import ComponentType from '../../../../../models/ComponentType';
 import { COMPONENTS_NAMES } from '../../../../../locales/lang-constants';
 import ComponentFactory from '../../../../../models/factories/ComponentFactory';
-import { LinkValidatorI } from '../../LinkValidator';
+import { LinkValidatorI } from '../../../LinkValidator';
 
 type NodeInterfaceFactoryProps = {
     factory: ComponentFactory
