+++ src/components/ComponentDiagram/elements/Port/PortWidget/Port.module.scss
@@ -1,4 +1,4 @@
-
+@import "~animate.css";
 
 $border-color: #cacaca;
 $dot-size: 14px;
